requirements.txt
aniso8601==7.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-ckeditor==5.7.1
django-easy-select2==1.5.6
django-graphql-jwt==0.2.2
django-import-export==2.3.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
promise==2.2.1
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
